Update paths after moving icons into sub-folder
