+++ src/components/Sidebar/Sidebar.Link.jsx
@@ -1,7 +1,7 @@
 import clsx from "clsx";
 import PropTypes from "prop-types";
 import React from "react";
-import { ChevronRightIcon } from "../../icons/ChevronRightIcon";
+import { ChevronRightIcon } from "../../icons/";
 import { Dot } from "../Dot/Dot";
 
 export const SidebarLink = ({ isActive = false, icon: Icon, children, isSubMenuItem, ...rest }) => {
