+++ src/helpers/flash.jsx
@@ -2,7 +2,7 @@ import { Transition } from "@headlessui/react";
 import clsx from "clsx";
 import React, { Fragment } from "react";
 import toast from "react-hot-toast";
-import { CloseIcon } from "../icons/CloseIcon";
+import { CloseIcon } from "../icons/";
 
 const colors = {
     primary: "bg-primary",
