+++ src/icons/index.js
@@ -20,8 +20,8 @@ export { BellIcon } from "./src/BellIcon";
 export { BookIcon } from "./src/BookIcon";
 export { BookmarkIcon } from "./src/BookmarkIcon";
 export { BoxIcon } from "./src/BoxIcon";
-export { BugIcon } from "./src/BugIcon";
 export { BriefcaseIcon } from "./src/BriefcaseIcon";
+export { BugIcon } from "./src/BugIcon";
 export { ButtonCodeIcon } from "./src/ButtonCodeIcon";
 export { CakeIcon } from "./src/CakeIcon";
 export { CalendarDayIcon } from "./src/CalendarDayIcon";
@@ -50,6 +50,7 @@ export { CircleDotIcon } from "./src/CircleDotIcon";
 export { CircleInfinityIcon } from "./src/CircleInfinityIcon";
 export { CircleInfoIcon } from "./src/CircleInfoIcon";
 export { CircleKeyIcon } from "./src/CircleKeyIcon";
+export { CircleNotch } from "./src/CircleNotch";
 export { CirclePauseIcon } from "./src/CirclePauseIcon";
 export { CirclePlusIcon } from "./src/CirclePlusIcon";
 export { CirclePlusImage } from "./src/images/CirclePlusImage";
@@ -98,6 +99,7 @@ export { FacebookIcon } from "./src/Logos/FacebookIcon";
 export { FaceNeutralIcon } from "./src/FaceNeutralIcon";
 export { FaceSadIcon } from "./src/FaceSadIcon";
 export { FaceSmileIcon } from "./src/FaceSmileIcon";
+export { FeedbackLightBulbIcon } from "./src/FeedbackLightBulbIcon";
 export { FilterIcon } from "./src/FilterIcon";
 export { FlagIcon } from "./src/FlagIcon";
 export { FlexFee2Icon } from "./src/FlexFee2Icon";
@@ -117,7 +119,6 @@ export { InvoiceIcon } from "./src/InvoiceIcon";
 export { KeyIcon } from "./src/KeyIcon";
 export { KioskIcon } from "./src/KioskIcon";
 export { LightIcon } from "./src/LightIcon";
-export { FeedbackLightBulbIcon } from "./src/FeedbackLightBulbIcon";
 export { LightningIcon } from "./src/LightningIcon";
 export { LinkIcon } from "./src/LinkIcon";
 export { LockIcon } from "./src/LockIcon";
@@ -207,5 +208,8 @@ export { WifiIcon } from "./src/WifiIcon";
 export { WriteIcon } from "./src/WriteIcon";
 export { XIcon } from "./src/Logos/XIcon";
 export { XolaBotIcon } from "./src/XolaBotIcon";
+export { XolaLogo } from "./src/images/XolaLogo";
+export { XolaLogoCircle } from "./src/images/XolaLogoCircle";
+export { XolaLogoSimple } from "./src/images/XolaLogoSimple";
 export { XrayIcon } from "./src/XrayIcon";
 export { YelpIcon } from "./src/Logos/YelpIcon";
