+++ src/icons/src/CircleNotch.jsx
@@ -1,11 +1,9 @@
 import React from "react";
 
-const CircleNotch = (props) => {
+export const CircleNotch = (props) => {
     return (
         <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" fill="none" viewBox="0 0 14 14" {...props}>
             <path fill="currentColor" d="M14 7a7 7 0 10-7 7v-1.4A5.6 5.6 0 1112.6 7H14z" />
         </svg>
     );
 };
-
-export default CircleNotch;
